Block Editor: Move state logic inside 'BlockRenameModal'
